+++ src/scripts/example.ts
@@ -1,0 +1,16 @@
+// example.ts
+/**
+ * Example type
+ */
+export type MyType = {
+    foo: string;
+    bar: number;
+  };
+/**
+ * Example Interface ABC
+ * just text here
+ * and here too
+ */
+  export interface abc {
+    var1: string;
+  }